Add files via upload
The advertising dataset is a collection of data about advertising campaigns and their impact on sales. The dataset includes information on the following variables:

TV: The advertising budget for TV
Radio: The advertising budget for radio
Newspaper: The advertising budget for newspaper
Sales: The number of units sold
The dataset can be used to understand the impact of different advertising channels on sales. It can also be used to optimize advertising campaigns to maximize sales.

The advertising dataset is a valuable resource for businesses that want to improve their marketing campaigns. It can help businesses to understand what channels are most effective for reaching their target audience and to allocate their advertising budgets accordingly.